# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portal Manajemen Proyek API
  version: 1.0.0
  description: REST API untuk autentikasi, RBAC, CRUD proyek, tugas, dashboard, risiko, time tracking, file, dan komentar.
servers:
  - url: http://localhost:4000/api
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: Daftar proyek sesuai role user login
      responses:
        "200":
          description: Daftar proyek terfilter RBAC
    post:
      tags: [Projects]
      summary: Membuat proyek, hanya Project Manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Proyek dibuat
        "403":
          description: Role tidak diizinkan
  /projects/{id}:
    put:
      tags: [Projects]
      summary: Mengubah proyek milik Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Proyek diperbarui
    delete:
      tags: [Projects]
      summary: Menghapus proyek milik Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Proyek dihapus
...
  /milestones/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah milestone, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Milestone diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus milestone, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Milestone dihapus
...
  /project-links/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah quick link, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Link diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus quick link, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Link dihapus
...
  /project-files/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah metadata file, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: File diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus metadata file, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: File dihapus
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold, on_track, at_risk, delayed]
        client_id:
          type: integer
        pm_id:
          type: integer
        cover_image_url:
          type: string
          format: uri
...